Refactor test installment.
