+++ test/functional/time_bookings_controller_test.rb
@@ -1,4 +1,4 @@
-require File.dirname(__FILE__) + '../../test_helper'
+require File.dirname(__FILE__) + '../../minitest_helper'
 
 
 class TimeBookingsControllerTest < ActionController::TestCase
