+++ test/functional/time_logs_controller_test.rb
@@ -1,4 +1,4 @@
-require File.dirname(__FILE__) + '../../test_helper'
+require File.dirname(__FILE__) + '../../minitest_helper'
 
 
 class TimeLogsControllerTest < ActionController::TestCase
@@ -103,7 +103,6 @@ class TimeLogsControllerTest < ActionController::TestCase
         assert_response 302, "on update TL"
         assert_equal(I18n.t(:tt_error_not_allowed_to_change_foreign_logs), flash[:error], "show error-message")
         tl = TimeLog.where(:id => 2).first
-        assert_equal(local_time("2012-10-25 8:47:00"), tl.started_on.to_time.localtime.strftime("%H:%M"), "not updated TL-time")
         assert_equal(local_datetime("2012-10-25 8:47:00"), tl.started_on, "not updated TL-date")
       end
     end
@@ -146,7 +145,6 @@ class TimeLogsControllerTest < ActionController::TestCase
         assert_response 302, "on update TL"
         assert_equal(I18n.t(:tt_update_log_success), flash[:notice] || flash[:error], "show flash-message")
         tl = TimeLog.where(:id => 1).first
-        assert_equal(local_time("2012-10-23 12:42:00"), tl.started_on.to_time.localtime.strftime("%H:%M"), "not updated TL-time")
         assert_equal(local_datetime("2012-10-23 12:42:00"), tl.started_on, "not updated TL-date")
       end
 
@@ -167,7 +165,6 @@ class TimeLogsControllerTest < ActionController::TestCase
         assert_response 302, "on update TL"
         assert_equal(I18n.t(:tt_error_not_allowed_to_change_foreign_logs), flash[:error], "show error-message")
         tl = TimeLog.where(:id => 2).first
-        assert_equal(local_time("2012-10-25 8:47:00"), tl.started_on.to_time.localtime.strftime("%H:%M"), "not updated TL-time")
         assert_equal(local_datetime("2012-10-25 8:47:00"), tl.started_on, "not updated TL-date")
       end
     end
@@ -210,7 +207,6 @@ class TimeLogsControllerTest < ActionController::TestCase
         assert_response 302, "on update TL"
         assert_equal(I18n.t(:tt_update_log_success), flash[:notice] || flash[:error], "show flash-message")
         tl = TimeLog.where(:id => 1).first
-        assert_equal(local_time("2012-10-23 12:42:00"), tl.started_on.to_time.localtime.strftime("%H:%M"), "not updated TL-time")
         assert_equal(local_datetime("2012-10-23 12:42:00"), tl.started_on, "not updated TL-date")
       end
 
@@ -231,7 +227,6 @@ class TimeLogsControllerTest < ActionController::TestCase
         assert_response 302, "on update TL"
         assert_equal(I18n.t(:tt_update_log_success), flash[:notice] || flash[:error], "show flash-message")
         tl = TimeLog.where(:id => 2).first
-        assert_equal(local_time("2012-10-23 12:42:00"), tl.started_on.to_time.localtime.strftime("%H:%M"), "not updated TL-time")
         assert_equal(local_datetime("2012-10-23 12:42:00"), tl.started_on, "not updated TL-date")
       end
     end
@@ -316,7 +311,6 @@ class TimeLogsControllerTest < ActionController::TestCase
         assert_response 302, "on update TL"
         assert_equal(I18n.t(:tt_error_not_allowed_to_change_logs), flash[:error], "show error-message")
         tl = TimeLog.where(:id => 1).first
-        assert_equal(local_time("2012-10-25 23:47:00"), tl.started_on.to_time.localtime.strftime("%H:%M"), "not updated TL-time")
         assert_equal(local_datetime("2012-10-25 23:47:00"), tl.started_on, "not updated TL-date")
       end
 
@@ -337,7 +331,6 @@ class TimeLogsControllerTest < ActionController::TestCase
         assert_response 302, "on update TL"
         assert_equal(I18n.t(:tt_error_not_allowed_to_change_foreign_logs), flash[:error], "show error-message")
         tl = TimeLog.where(:id => 2).first
-        assert_equal(local_time("2012-10-25 8:47:00"), tl.started_on.to_time.localtime.strftime("%H:%M"), "not updated TL-time")
         assert_equal(local_datetime("2012-10-25 8:47:00"), tl.started_on, "not updated TL-date")
       end
     end
@@ -402,7 +395,6 @@ class TimeLogsControllerTest < ActionController::TestCase
         assert_response 302, "on update TL"
         assert_equal(I18n.t(:tt_error_not_allowed_to_change_logs), flash[:error], "show error-message")
         tl = TimeLog.where(:id => 1).first
-        assert_equal(local_time("2012-10-25 23:47:00"), tl.started_on.to_time.localtime.strftime("%H:%M"), "not updated TL-time")
         assert_equal(local_datetime("2012-10-25 23:47:00"), tl.started_on, "not updated TL-date")
       end
 
@@ -423,7 +415,6 @@ class TimeLogsControllerTest < ActionController::TestCase
         assert_response 302, "on update TL"
         assert_equal(I18n.t(:tt_error_not_allowed_to_change_foreign_logs), flash[:error], "show error-message")
         tl = TimeLog.where(:id => 2).first
-        assert_equal(local_time("2012-10-25 8:47:00"), tl.started_on.to_time.localtime.strftime("%H:%M"), "not updated TL-time")
         assert_equal(local_datetime("2012-10-25 8:47:00"), tl.started_on, "not updated TL-date")
       end
     end
